PheggMod "MegaPatch" :3
Added random scaling and infection systems (along with a custom PlayerEffect for infected players). PlayerHurt and PlayerDeath events have been moved to after the data is handled, meaning that to get the original data you now need to use the cache for both events. And a general cleanup of the code (including moving everything to referencehubs)